requirements.txt
aiofiles==23.2.1
aiogram==2.14
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.12.1
annotated-doc==0.0.3
annotated-types==0.6.0
anyio==3.7.1
apple-compress==0.2.3
APScheduler==3.6.3
arrow==1.3.0
asn1==2.7.1
asttokens==3.0.0
attrs==23.2.0
Babel==2.14.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
bpylist2==4.1.1
Brotli==1.1.0
cached-property==2.0.1
cachetools==4.2.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
construct==2.10.70
cryptography==44.0.0
daemonize==2.5.0
decorator==5.1.1
developer_disk_image==0.0.4
dnspython==2.7.0
ecdsa==0.19.1
editor==1.6.6
email_validator==2.2.0
enum-compat==0.0.3
executing==2.1.0
fastapi==0.104.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
Flask==3.0.3
frozenlist==1.4.1
gpxpy==1.5.0
h11==0.14.0
hexdump==3.3
httpcore==1.0.2
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
ifaddr==0.2.0
inquirer3==0.6.1
ipsw-parser==1.3.9
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
la-panic==0.5.0
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mutagen==1.47.0
nest-asyncio==1.6.0
opack==0.1.0
orjson==3.10.14
packaging==24.2
parameter-decorators==0.0.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
plumbum==1.9.0
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pycrashreport==1.2.5
pycryptodome==3.21.0
pycryptodomex==3.20.0
pydantic==2.12.3
pydantic-extra-types==2.10.1
pydantic-settings==2.1.0
pydantic_core==2.41.4
Pygments==2.19.1
pygnuutils==0.1.1
pyimg4==0.8.6
pykdebugparser==1.2.4
pylzss==0.3.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==13.7
pytun-pmd3==2.0.10
pytz==2023.3.post1
pyusb==1.3.1
PyYAML==6.0.2
qh3==1.3.0
qrcode==7.4.2
readchar==4.2.1
remotezip==0.12.3
remotezip2==0.0.2
requests==2.31.0
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
runs==1.2.2
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
srptools==1.0.1
stack-data==0.6.3
starlette==0.27.0
termcolor==2.5.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
xmod==1.8.1
xonsh==0.19.0
yarl==1.9.4
yt-dlp==2024.5.27
zeroconf==0.139.0
